+++ momoko/__init__.py
@@ -1,3 +1,12 @@
-#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+"""
+    momoko
+    ~~~~~~
+
+    Asynchronous Psycopg wrapper for Tornado.
+
+    :copyright: (c) 2011 by Frank Smit.
+    :license: MIT, see LICENSE for more details.
+"""
 
 from momoko import *