+++ docs/news.rst
@@ -1,9 +1,11 @@
 News/Changelog
-=========================
+==============
+
 
 0.2.0 (2011-??-??)
 ------------------
 
+* Added a wrapper class, ``Momoko``, for ``Pool``, ``BatchQuery`` and ``QueryChain``.
 * Added the ``QueryChain`` class for executing a chain of queries (and callables)
   in a certain order.
 * Added the ``BatchQuery`` class for executing batches of queries at the same time.